Extracted common tuning meter component code to the library module.
